feat: Buno everywhere (#4)
* feat: Allow commands to be run as user app

* feat: Declare if app is installed in guild or not

* fix: remove useless console.log

* feat: Begin working on Buno Everywhere

* feat: Buno everywhere update

* fix: Types

* feat: Add whitelist to Buno Everywhere

* fix: Settings

* chore: Add GH Actions to dependabot

* chore(deps): bump actions/cache from 4.0.2 to 4.2.3 (#5) [no ci]

Bumps [actions/cache](https://github.com/actions/cache) from 4.0.2 to 4.2.3.
- [Release notes](https://github.com/actions/cache/releases)
- [Changelog](https://github.com/actions/cache/blob/main/RELEASES.md)
- [Commits](https://github.com/actions/cache/compare/0c45773b623bea8c8e75f6c82b208c3cf94ea4f9...5a3ec84eff668545956fd18022155c47e93e2684)

---
updated-dependencies:
- dependency-name: actions/cache
  dependency-version: 4.2.3
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* chore(deps): bump pnpm/action-setup from 3.0.0 to 4.1.0 (#7) [no ci]

Bumps [pnpm/action-setup](https://github.com/pnpm/action-setup) from 3.0.0 to 4.1.0.
- [Release notes](https://github.com/pnpm/action-setup/releases)
- [Commits](https://github.com/pnpm/action-setup/compare/a3252b78c470c02df07e9d59298aecedc3ccdd6d...a7487c7e89a18df4991f7f222e4898a00d66ddda)

---
updated-dependencies:
- dependency-name: pnpm/action-setup
  dependency-version: 4.1.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* chore(deps): bump actions/checkout from 4.1.1 to 4.2.2 (#6) [no ci]

Bumps [actions/checkout](https://github.com/actions/checkout) from 4.1.1 to 4.2.2.
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/b4ffde65f46336ab88eb53be808477a3936bae11...11bd71901bbe5b1630ceea73d27597364c9af683)

---
updated-dependencies:
- dependency-name: actions/checkout
  dependency-version: 4.2.2
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* chore(deps): bump actions/setup-node from 4.0.2 to 4.3.0 (#8) [no ci]

Bumps [actions/setup-node](https://github.com/actions/setup-node) from 4.0.2 to 4.3.0.
- [Release notes](https://github.com/actions/setup-node/releases)
- [Commits](https://github.com/actions/setup-node/compare/60edb5dd545a775178f52524783378180af0d1f8...cdca7365b2dadb8aad0a33bc7601856ffabcc48e)

---
updated-dependencies:
- dependency-name: actions/setup-node
  dependency-version: 4.3.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* fix: Version number in CI

* feat: Allow commands to be run as user app

* feat: Declare if app is installed in guild or not

* fix: remove useless console.log

* feat: Begin working on Buno Everywhere

* feat: Buno everywhere update

* fix: Types

* feat: Add whitelist to Buno Everywhere

* fix: Settings

---------

+++ src/interactions/buttons/confirmLeaveGame.ts
@@ -1,5 +1,5 @@
 
-import { MessageFlags, TextChannel } from "discord.js";
+import { MessageEditOptions, MessageFlags, TextChannel } from "discord.js";
 import { t } from "i18next";
 
 import runningGameMessage from "../../components/runningGameMessage.js";
@@ -37,10 +37,20 @@ export const b: button = {
         game.playersWhoLeft.push(interaction.user.id);
         timeouts.delete(game.channelId);
         timeouts.set(game.channelId, () => onTimeout(client, game, game.currentPlayer), game.settings.timeoutDuration * 1000);
-        await interaction.channel?.messages.cache.get(game.messageId)?.delete();
-        await (interaction.channel as TextChannel).send(t("strings:game.left", { name: await getUsername(client, game.guildId, interaction.user.id), lng }));
+        const leaveMessage = t("strings:game.left", { name: await getUsername(client, game.guildId, interaction.user.id, !game.guildApp), lng });
+        if (game.guildApp) {
+            await interaction.channel?.messages.cache.get(game.messageId)?.delete();
+            await (interaction.channel as TextChannel).send(leaveMessage);
+        }
+        else game.previousActions.push(leaveMessage);
         if ((game._modified && game.players.length === 0) || (!game._modified && game.players.length === 1)) return endGame(game, interaction.client, "notEnoughPeople");
-        const msg = await (interaction.channel as TextChannel).send(await runningGameMessage(client, game));
-        game.messageId = msg.id;
+        if (game.guildApp) {
+            const msg = await (interaction.channel as TextChannel).send(await runningGameMessage(client, game));
+            game.messageId = msg.id;
+        }
+        else await interaction.editReply({
+            message: game.messageId,
+            ...await runningGameMessage(client, game) as MessageEditOptions
+        });
     }
 };
